+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### __Breaking Changes__
 
 ### Added
+- App: Added automatic device mode switch between Desktop and Mobile when window size changes.
 
 ### Changed